+++ CHANGES.rst
@@ -10,6 +10,7 @@ Development
 - [ui] Reorder comfort form fields again
 - [ui] Nail Javascript module dependencies to exact versions
 - [mw] Fix error reporting and logging issues
+- [ui] Hide previous/next drawing navigation buttons when displaying a single drawing
 
 
 2018-10-08 0.153.0